delay/requirements.txt
accelerate==0.29.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diffusers==0.29.2
dynamicprompts==0.31.0
filelock==3.13.3
Flask==3.0.3
fsspec==2024.3.1
huggingface-hub==0.24.1
idna==3.6
importlib_metadata==8.1.0
inquirerpy==0.3.4
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pfzy==0.3.4
pillow==10.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
pyparsing==3.1.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scipy==1.14.0
Send2Trash==1.8.3
sympy==1.12
tbb==2021.12.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
transformers @ git+https://github.com/huggingface/transformers@c85510f958e6955d88ea1bafb4f320074bfbd0c1
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.19.2
